highscore

+++ src/engine/HighScore.java
@@ -2,6 +2,7 @@ package engine;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.FileWriter;
@@ -14,6 +15,10 @@ public class HighScore {
 	static public int readHighScore() {
 		BufferedReader in;
 		try {
+			File f = new File(filename);
+			if (!f.exists()) {
+				return 0;
+			}
 			in = new BufferedReader(new FileReader(filename));
 			String line;
 			line = in.readLine();
@@ -36,6 +41,13 @@ public class HighScore {
 		BufferedReader in;
 		BufferedWriter out;
 		try {
+			File f = new File(filename);
+			if (!f.exists()) {
+				out = new BufferedWriter(new FileWriter(filename));	
+				out.write(score);
+				out.close();
+				return;
+			}			
 			in = new BufferedReader(new FileReader(filename));
 			out = new BufferedWriter(new FileWriter(filename));
 			String line;
